requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0

# LLM Providers
openai==1.12.0
requests==2.31.0

# Vector Database
qdrant-client==1.7.3

# Finance Data
yfinance==0.2.36

# Utilities
python-dotenv==1.0.1
